Move similar logic to mixins.
