+++ admin/src/pages/media.vue
@@ -38,7 +38,8 @@
 </template>
 
 <script>
-  /* global EventHub */
+  import CreateItemFromImages from '../mixins/CreateItemFromImages'
+  import SnackBar from '../mixins/SnackBar'
 
   export default {
     name: 'Media',
@@ -50,53 +51,6 @@
       // DashboardCoreView: () => import('../components/core/View'),
     },
 
-    data: () => ({
-      snackbar: false,
-      snackbarText: '',
-      snackbarColor: '',
-    }),
-
-    created () {
-      EventHub.listen('show-snackbar', (options) => this.showSnackbar(options.color, options.text))
-      EventHub.listen('MediaManager-create-new-item', (images) => this.createItemFromImages(images))
-    },
-
-    beforeDestroy () {
-      EventHub.removeListenersFrom([
-        'show-snackbar',
-        'MediaManager-create-new-item',
-      ])
-    },
-
-    methods: {
-      showSnackbar (color, text) {
-        // TODO: after switching to a real SPA move this into the header component
-        this.snackbarColor = color
-        this.snackbarText = text
-        this.snackbar = true
-      },
-
-      showSnackbarSuccess (text) {
-        // TODO: after switching to a real SPA move this into the header component
-        this.showSnackbar('success', text)
-      },
-
-      showSnackbarError (text = 'An error occurred') {
-        // TODO: after switching to a real SPA move this into the header component
-        this.showSnackbar('error', text)
-      },
-
-      async createItemFromImages (images) {
-        try {
-          const { data } = await this.$http.post('/api/items?project=catalogue', {
-            images: images,
-          })
-
-          window.location.href = `/staff/items/${data.id}`
-        } catch (e) {
-          this.showSnackbarError()
-        }
-      },
-    },
+    mixins: [CreateItemFromImages, SnackBar],
   }
 </script>
